requirements.txt
flask==2.3.2
flask-mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.2.0